feat: Login_record, fix: migration, add: models & controller edit_record

+++ backend/routes/web.php
@@ -192,13 +192,14 @@ Route::middleware(['auth'])->group(function () {
     Route::put('/page/about/edit/{id}', [PagesController::class, 'editAbout'])->name('edit-about');
     Route::put('/page/service/edit/{id}', [PagesController::class, 'editService'])->name('page-settings.update');
 
-    //! To do List
-    Route::get('/todolist', [ToDoListController::class, 'getToDoList'])->name('todolist.index');
-    Route::get('/todolist/create', [ToDoListController::class, 'addToDoList'])->name('adminpanel.todolist.create');
-    Route::post('/todolist/store', [ToDoListController::class, 'storeToDoList'])->name('adminpanel.todolist.store');
-    Route::put('/todolist/{todolist}', [ToDoListController::class, 'updateToDoList'])->name('adminpanel.todolist.update');
-    Route::get('/todolist/{todolist}/edit', [ToDoListController::class, 'editToDoList'])->name('adminpanel.todolist.edit');
-    Route::delete('/todolist/{todolist}', [ToDoListController::class , 'deleteToDoList'])->name('adminpanel.todolist.destroy');
+    //! TODOLIST
+    Route::get('adminpanel/todolist', [ToDoListController::class, 'getToDoList'])->name('todolist.index');
+    Route::get('adminpanel/todolist/create', [ToDoListController::class, 'addToDoList'])->name('adminpanel.todolist.create');
+    Route::post('/adminpanel/todolist/store', [ToDoListController::class, 'storeToDoList'])->name('adminpanel.todolist.store');
+    Route::put('/adminpanel/todolist/{todolist}', [ToDoListController::class, 'updateToDoList'])->name('adminpanel.todolist.update');
+    Route::get('/adminpanel/todolist/{todolist}/edit', [ToDoListController::class, 'editToDoList'])->name('adminpanel.todolist.edit');
+    Route::delete('/adminpanel/todolist/{todolist}', [ToDoListController::class , 'deleteToDoList'])->name('adminpanel.todolist.destroy');
+
 });
 
 // hanya untuk user dengan role superadmin
